app/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.7.0
appier==1.34.5
async-lru==2.0.4
asyncpg==0.29.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
google==3.0.0
google-api==0.1.12
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.2.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
Telethon==1.38.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
yfinance==0.2.4
todoist-api-python==2.1.7

# pip freeze > requirements.txt
# pip install -r requirements.txt
